refactor(core): move parseWikilink into wikilink.ts
